code refactor to restaurant subdirectory
