+++ food-ordering/app/app/urls.py
@@ -24,6 +24,7 @@ from django.conf.urls.static import static
 
 urlpatterns = [
     path('', include('starting.urls')),
+    path('', include('restaurant.urls')),
     path('', include('django.contrib.auth.urls')),
     path('user/', include('user.urls')),
     path('admin/', admin.site.urls),
