+++ food-ordering/app/restaurant/admin.py
@@ -1,3 +1,7 @@
 from django.contrib import admin
+from .models import Food, Reservation
 
 # Register your models here.
+
+admin.site.register(Food)
+admin.site.register(Reservation)
