+++ food-ordering/app/restaurant/templates/restaurant/reservation.html
@@ -1,0 +1,69 @@
+{% extends "master.html" %}
+{% block title %}Book A Table{% endblock%}
+{% block head %}{% endblock %}
+{% block body %}
+    <div class="container-xxl bg-white p-0">
+        <!-- Spinner Start -->
+        <!-- <div id="spinner" class="show bg-white position-fixed translate-middle w-100 vh-100 top-50 start-50 d-flex align-items-center justify-content-center">
+      <div class="spinner-border text-primary" style="width: 3rem; height: 3rem" role="status">
+        <span class="sr-only">Loading...</span>
+      </div>
+    </div> -->
+        <!-- Spinner End -->
+        <!-- Reservation Start -->
+        <div class="container-xxl py-5 px-0 wow fadeInUp" data-wow-delay="0.1s">
+            <div class="row g-0 pt-4">
+                <div class="col-md-6">
+                    <div class="video">
+                        <button type="button"
+                                class="btn-play"
+                                data-bs-toggle="modal"
+                                data-src="https://www.youtube.com/embed/DWRcNpR6Kdc"
+                                data-bs-target="#videoModal">
+                            <span>0</span>
+                        </button>
+                    </div>
+                </div>
+                <div class="col-md-6 bg-dark d-flex align-items-center justify-content-center row">
+                    <div class="p-5 wow fadeInUp" data-wow-delay="0.2s">
+                        <h1 class="section-title ff-secondary text-start text-primary fw-normal">Reservation</h1>
+                        <h1 class="text-white mb-4">Book A Table Online</h1>
+                        {% include "restaurant/reservationForm.html" %}
+                
+        </div>
+    </div>
+</div>
+</div>
+<div class="modal fade"
+     id="videoModal"
+     tabindex="-1"
+     aria-labelledby="exampleModalLabel"
+     aria-hidden="true">
+    <div class="modal-dialog">
+        <div class="modal-content rounded-0">
+            <div class="modal-header">
+                <h5 class="modal-title" id="exampleModalLabel">Youtube Video</h5>
+                <button type="button"
+                        class="btn-close"
+                        data-bs-dismiss="modal"
+                        aria-label="Close"></button>
+            </div>
+            <div class="modal-body">
+                <!-- 16:9 aspect ratio -->
+                <div class="ratio ratio-16x9">
+                    <iframe class="embed-responsive-item"
+                            src=""
+                            id="video"
+                            allowfullscreen
+                            allowscriptaccess="always"
+                            allow="autoplay"></iframe>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+<!-- Reservation End -->
+<!-- Back to Top -->
+<a href="#" class="btn btn-lg btn-primary btn-lg-square back-to-top"><i class="bi bi-arrow-up"></i></a>
+</div>
+{% endblock %}
