+++ food-ordering/app/restaurant/templates/restaurant/reservationForm.html
@@ -1,0 +1,29 @@
+{% load crispy_forms_tags %}
+<form action="" method="post" autocomplete="off">
+    {% csrf_token %}
+    {{ reservationForm|crispy }}
+    <div class="col-12">
+                    <button type="submit"
+                            class="btn btn-primary w-100 py-3"
+                            name="Book Now"
+                            value="Book Now">Book Now</button>
+                </div>
+</form>
+
+
+
+<form action="" method="post" autocomplete="off">
+    {% csrf_token %}
+    <div class="row">
+        <div class="fieldWrapper">
+            {{ reservationForm.customer_name.as_field_group }}
+        </div>
+    </div>
+
+    <div class="col-12">
+                    <button type="submit"
+                            class="btn btn-primary w-100 py-3"
+                            name="Book Now"
+                            value="Book Now">Book Now</button>
+                </div>
+</form>
