+++ food-ordering/app/starting/admin.py
@@ -1,7 +1,3 @@
 from django.contrib import admin
-from .models import Food, Reservation
 
 # Register your models here.
-
-admin.site.register(Food)
-admin.site.register(Reservation)
